+++ core-ng/src/main/java/core/framework/api/module/DBConfig.java
@@ -8,8 +8,6 @@ import core.framework.api.util.Strings;
 import core.framework.api.util.Types;
 import core.framework.impl.db.DatabaseImpl;
 import core.framework.impl.module.ModuleContext;
-import core.framework.impl.resource.RefreshPoolJob;
-import core.framework.impl.scheduler.FixedRateTrigger;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -32,10 +30,14 @@ public final class DBConfig {
             database = context.beanFactory.bean(Database.class, name);
         } else {
             database = new DatabaseImpl();
+            database.pool.name("db" + (name == null ? "" : "-" + name));
+
             context.shutdownHook.add(database::close);
-            String poolName = "db" + (name == null ? "" : "-" + name);
-            database.pool.name(poolName);
-            context.scheduler().addTrigger(new FixedRateTrigger("refresh-" + poolName + "-pool", new RefreshPoolJob(database.pool), Duration.ofMinutes(15)));
+
+            if (!context.isTest()) {
+                context.backgroundTask().scheduleWithFixedDelay(database.pool::refresh, Duration.ofMinutes(30));
+            }
+
             context.beanFactory.bind(Database.class, name, database);
         }
     }