+++ core-ng/src/main/java/core/framework/api/module/QueueConfig.java
@@ -8,8 +8,6 @@ import core.framework.impl.queue.RabbitMQ;
 import core.framework.impl.queue.RabbitMQEndpoint;
 import core.framework.impl.queue.RabbitMQListener;
 import core.framework.impl.queue.RabbitMQPublisher;
-import core.framework.impl.resource.RefreshPoolJob;
-import core.framework.impl.scheduler.FixedRateTrigger;
 
 import java.time.Duration;
 
@@ -27,7 +25,7 @@ public final class QueueConfig {
         if (context.queueManager.rabbitMQ == null) {
             context.queueManager.rabbitMQ = new RabbitMQ();
             if (!context.isTest()) {
-                context.scheduler().addTrigger(new FixedRateTrigger("refresh-rabbitmq-pool", new RefreshPoolJob(context.queueManager.rabbitMQ.pool), Duration.ofMinutes(5)));
+                context.backgroundTask().scheduleWithFixedDelay(context.queueManager.rabbitMQ.pool::refresh, Duration.ofMinutes(5));
                 context.shutdownHook.add(context.queueManager.rabbitMQ::close);
             }
         }