+++ core-ng/src/main/java/core/framework/api/module/SessionConfig.java
@@ -2,9 +2,6 @@ package core.framework.api.module;
 
 import core.framework.impl.module.ModuleContext;
 import core.framework.impl.redis.RedisImpl;
-import core.framework.impl.resource.RefreshPoolJob;
-import core.framework.impl.scheduler.FixedRateTrigger;
-import core.framework.impl.web.session.CleanupLocalSessionStoreJob;
 import core.framework.impl.web.session.LocalSessionStore;
 import core.framework.impl.web.session.RedisSessionStore;
 import org.slf4j.Logger;
@@ -30,7 +27,9 @@ public final class SessionConfig {
     public void local() {
         logger.info("create local session provider");
         LocalSessionStore sessionStore = new LocalSessionStore();
-        context.scheduler().addTrigger(new FixedRateTrigger("cleanup-local-session", new CleanupLocalSessionStoreJob(sessionStore), Duration.ofMinutes(30)));
+        if (!context.isTest()) {
+            context.backgroundTask().scheduleWithFixedDelay(sessionStore::cleanup, Duration.ofMinutes(30));
+        }
         context.httpServer.siteManager.sessionManager.sessionStore(sessionStore);
     }
 
@@ -44,7 +43,7 @@ public final class SessionConfig {
             RedisImpl redis = new RedisImpl();
             redis.host(host);
             redis.pool.name("redis-session");
-            context.scheduler().addTrigger(new FixedRateTrigger("refresh-redis-session-pool", new RefreshPoolJob(redis.pool), Duration.ofMinutes(5)));
+            context.backgroundTask().scheduleWithFixedDelay(redis.pool::refresh, Duration.ofMinutes(5));
 
             context.shutdownHook.add(redis::close);
             context.httpServer.siteManager.sessionManager.sessionStore(new RedisSessionStore(redis));